requirements.txt
Pillow
django-cors-headers
